+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # main
 - New class:        In Condition: Matrix to evaluate matrix elements.
+- New class:        GnssParametrizationTransmitterQzss: QZSS constellation support in GNSS processing.
 - New program:      InstrumentAccelerometer2ThermosphericDensity: Estimate neutral density from accelerometer data.
 
 # Release 2021-02-02